requirements_new.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# File processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-magic>=0.4.27
python-magic-bin>=0.4.14  # For Windows

# AI and NLP
openai>=1.0.0
transformers>=4.30.0
reportlab>=4.0.0

# AI and NLP

# Security
cryptography>=41.0.0
hashlib-compat>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0

# Optional: For better performance
accelerate>=0.20.0
sentencepiece>=0.1.99

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0